requirements.txt
pandas
PyPDF2
streamlit
plaid-python
